FIX: 코드 리팩토링 및 변경된 기획사항 반영
